added navigation tests.
