+++ examples/app-basic/src/constants.ts
@@ -1,14 +1,12 @@
 import { NavModelItem } from '@grafana/data';
 import pluginJson from './plugin.json';
 
-
 export const PLUGIN_BASE_URL = `/a/${pluginJson.id}`;
 
 export enum ROUTES {
   One = 'one',
   Two = 'two',
   Three = 'three',
-  Four = 'four',
 }
 
 export const NAVIGATION_TITLE = 'Basic App Plugin';
