+++ examples/app-basic/src/pages/PageThree.tsx
@@ -1,46 +1,41 @@
 import React from 'react';
 import { css } from '@emotion/css';
-import { useParams, Link } from 'react-router-dom';
-import { GrafanaTheme2 } from '@grafana/data';
-import { useStyles2 } from '@grafana/ui';
-import { prefixRoute } from '../utils/utils.routing';
+import { GrafanaTheme2, PageLayoutType } from '@grafana/data';
+import { LinkButton, useStyles2 } from '@grafana/ui';
 import { ROUTES } from '../constants';
+import { prefixRoute } from '../utils/utils.routing';
 import { PluginPage } from '@grafana/runtime';
 
 export function PageThree() {
   const s = useStyles2(getStyles);
-  const { id } = useParams<{ id: string }>();
 
   return (
-    <PluginPage>
-      <div>
-        This is page three.
-        <br />
-        <br />
-        {/* The ID parameter is set */}
-        {id && (
-          <>
-            <strong>ID:</strong> {id}
-          </>
-        )}
-        {/* No ID parameter */}
-        {!id && (
-          <>
-            <strong>No id parameter is set in the URL.</strong> <br />
-            Try the following link: <br />
-            <Link className={s.link} to={prefixRoute(`${ROUTES.Three}/123456789`)}>
-              {prefixRoute(`${ROUTES.Three}/123456789`)}
-            </Link>
-          </>
-        )}
+    <PluginPage layout={PageLayoutType.Canvas}>
+      <div className={s.page}>
+        <div className={s.container}>
+          <LinkButton icon="arrow-left" href={prefixRoute(ROUTES.One)}>
+            Back
+          </LinkButton>
+          <div className={s.content}>This is a full-width page without a navigation bar.</div>
+        </div>
       </div>
     </PluginPage>
   );
 }
 
 const getStyles = (theme: GrafanaTheme2) => ({
-  link: css`
-    color: ${theme.colors.text.link};
-    text-decoration: underline;
+  page: css`
+    padding: ${theme.spacing(3)};
+    background-color: ${theme.colors.background.secondary};
+    display: flex;
+    justify-content: center;
+  `,
+  container: css`
+    width: 900px;
+    max-width: 100%;
+    min-height: 500px;
+  `,
+  content: css`
+    margin-top: ${theme.spacing(6)};
   `,
 });
